+++ demo/DemoV2/Client/UserDemoClient.generated.cs
@@ -335,9 +335,9 @@ namespace DemoV2.Client
             if(filter != null)
                 queryBuilder.Add("filter", filter.Value());
             if(include != null)
-                queryBuilder.Add("include", include.AsStringList());
+                queryBuilder.Add("include[]", include.AsStringList());
             if(flags != null)
-                queryBuilder.Add("flags", Invariant($"{flags}"));
+                queryBuilder.Add("flags[]", Invariant($"{flags}"));
             if(exclude != null)
                 queryBuilder.Add("exclude", exclude.AsStringList());
             uriBuilder.Query = queryBuilder.ToString() ?? string.Empty;